# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DocSecur API",
        "description": "API REST pour la plateforme de gestion médicale DocSecur. Authentification via Bearer Token (Laravel Passport).",
        "contact": {
            "email": "support@docsecur.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/medecin/prescriptions": {
            "get": {
                "tags": [
                    "Médecin - Prescriptions"
                ],
                "summary": "Lister les prescriptions du médecin",
                "description": "Retourne la liste paginée des prescriptions émises par le médecin connecté.",
                "operationId": "8403b134288ff7c5c2236bcb15e6a995",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des prescriptions paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Médecin - Prescriptions"
                ],
                "summary": "Créer une prescription",
                "description": "Crée une ordonnance avec la liste des médicaments prescrits. Le numéro d'ordonnance est généré automatiquement.",
                "operationId": "a3f5954e3454ea5169f1cb0a7981c161",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "consultation_id",
                                    "medicaments"
                                ],
                                "properties": {
                                    "consultation_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Prendre avec de la nourriture"
                                    },
                                    "pharmacie_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "medicaments": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "medicament_id",
                                                "posologie"
                                            ],
                                            "properties": {
                                                "medicament_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "posologie": {
                                                    "type": "string",
                                                    "example": "2 comprimés matin et soir"
                                                },
                                                "duree_traitement": {
                                                    "description": "Durée en jours",
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "quantite": {
                                                    "type": "integer",
                                                    "example": 14
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Prescription créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Prescription créée avec succès"
                                        },
                                        "prescription": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/medecin/prescriptions/{id}/envoyer-pharmacie": {
            "post": {
                "tags": [
                    "Médecin - Prescriptions"
                ],
                "summary": "Envoyer une ordonnance à la pharmacie",
                "description": "Associe une pharmacie à l'ordonnance et change son statut en 'envoyée'.",
                "operationId": "47e132ce9a1114f6d7e271a4d10a12d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la prescription",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pharmacie_id"
                                ],
                                "properties": {
                                    "pharmacie_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ordonnance envoyée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ordonnance envoyée à la pharmacie"
                                        },
                                        "prescription": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Prescription non trouvée"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/medecin/prescriptions/{id}": {
            "get": {
                "tags": [
                    "Médecin - Prescriptions"
                ],
                "summary": "Détails d'une prescription",
                "operationId": "93dabad5ca39af9a689f7738b165d7b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la prescription",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Prescription non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Médecin - Prescriptions"
                ],
                "summary": "Supprimer une prescription",
                "operationId": "6d905423062cddd9da5392ae17df4373",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prescription supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Prescription supprimée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Prescription non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}